# Contracts/openapi.snapshot.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "BackendApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/classroom-discussions/{id}/posts": {
      "post": {
        "tags": [
          "ClassroomDiscussions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomDiscussionPostRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomDiscussionPostRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClassroomDiscussionPostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ClassroomDiscussions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClassroomDiscussionPostDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/classroom-discussions/{id}/materials": {
      "get": {
        "tags": [
          "ClassroomDiscussions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassroomDiscussionPostDto": {
        "required": [
          "id",
          "classroomDiscussionId",
          "authorId",
          "content",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "classroomDiscussionId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateClassroomDiscussionPostRequest": {
        "required": [
          "content"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "MaterialDto": {
        "required": [
          "id",
          "title",
          "fileUrl",
          "subjectId",
          "clubId",
          "classroomDiscussionId",
          "staffGroupId",
          "uploadedBy",
          "uploadedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "subjectId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "clubId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "classroomDiscussionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "staffGroupId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "uploadedBy": {
            "type": "string",
            "format": "uuid"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}